🐛 fix: fix cyclic dependency between auth, project and organization modules
Resolves rapyuta-robotics/rapyuta_io#888
