+++ riocli/exceptions/__init__.py
@@ -56,3 +56,13 @@ class OperationNotSupported(Exception):
     def __init__(self, message="operation not supported"):
         self.message = message
         super().__init__(self.message)
+
+class OrganizationNotFound(Exception):
+    def __init__(self, message="organization not found"):
+        self.message = message
+        super().__init__(self.message)
+
+class ProjectNotFound(Exception):
+    def __init__(self, message="project not found"):
+        self.message = message
+        super().__init__(self.message)
